--- src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/payU/PayUPort.java
@@ -1,9 +1,0 @@
-package com.wojciechbarwinski.demo.epic_board_games_shop.payU;
-
-import com.wojciechbarwinski.demo.epic_board_games_shop.entities.Order;
-import com.wojciechbarwinski.demo.epic_board_games_shop.payU.dtos.PayUOrderResponseDTO;
-
-public interface PayUPort {
-
-    PayUOrderResponseDTO getPayUOrder(Order order);
-}
