+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/payments/PaymentDataDTO.java
@@ -1,0 +1,10 @@
+package com.wojciechbarwinski.demo.epic_board_games_shop.payments;
+
+
+import lombok.Data;
+
+@Data
+public class PaymentDataDTO {
+    private String orderId;
+    private String redirectUri;
+}
