+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/payments/PaymentPort.java
@@ -1,0 +1,8 @@
+package com.wojciechbarwinski.demo.epic_board_games_shop.payments;
+
+import com.wojciechbarwinski.demo.epic_board_games_shop.entities.Order;
+
+public interface PaymentPort {
+
+    PaymentDataDTO createPaymentData(Order order);
+}
