+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/payments/PaymentService.java
@@ -1,0 +1,18 @@
+package com.wojciechbarwinski.demo.epic_board_games_shop.payments;
+
+import com.wojciechbarwinski.demo.epic_board_games_shop.entities.Order;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class PaymentService {
+
+    private final PaymentPort paymentPort;
+
+        public PaymentDataDTO createPaymentData(Order order){
+
+            return paymentPort.createPaymentData(order);
+        }
+
+}
