+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/payments/payU/clients/PayUClientImpl.java
@@ -1,0 +1,24 @@
+package com.wojciechbarwinski.demo.epic_board_games_shop.payments.payU.clients;
+
+import com.wojciechbarwinski.demo.epic_board_games_shop.payments.payU.dtos.PayUTokenResponseDTO;
+import lombok.RequiredArgsConstructor;
+import okhttp3.Response;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class PayUClientImpl implements PayUClient {
+
+    private final PayUAuthClient payUAuthClient;
+    private final PayUOrderClient payUOrderClient;
+
+    @Override
+    public PayUTokenResponseDTO getAccessToken(String grantType, String clientId, String clientSecret) {
+        return payUAuthClient.getAccessToken(grantType, clientId, clientSecret);
+    }
+
+    @Override
+    public Response getPayUOrder(String token, String orderRequestJSON) {
+        return payUOrderClient.getNewPaymentOrder(token, orderRequestJSON);
+    }
+}
